lambda/add_two_numbers/requirements.txt
awslabs.mcp-lambda-handler==0.1.4